Added DatabaseFacade Extensions
Added DatabaseFacade extension method with an `IsFirebirdSql` method. By calling `dbContext.Database.IsFirebirdSql`, it will return `true` if the current database provider is FirebirdSQL, otherwise it will return false.
This is useful when writing code which is designed to be used with multiple different database providers, and specific logic is required for FirebirdSQL.