Changes for DNET-784: schema update for identity
Added a refactoring to FbSchema so the major version number is stored in the object. This then is used to conditionally configure the schema query for the Columns schema so it obtains rdb$relation_fields.rdb$identity_type.
In the returned datatable a new column IS_IDENTITY is added, type bool, which is set to true if the identity_type field in the resultset is 1.