Update FbDataReader.cs GetSchemaTable
Modified GetSchemaTable/Async to a single request to the FB-Server.
At the moment, in a loop is made a request for each field. If no transaction is specified, additional internal transaction is created.
So if you have N fields in the result, you made N requests to the server, perhaps including many transactions.
If a query is made with derived tables, the relation fo fields get empty, but the request is done.
So you can be faster with a single call and single transaction to the server and load all fields of involved relations.
The FetchSize can increase to the max to get all at once.
The final build for the SchemaTable can be done within one loop.
Also GetSchemaTable/Async and share the code for building teh schema.
And real, the GetSchemaTable for 100 columns is less then 100ms.